Fix: Increase p-value precision in MLR
The multiple linear regression (MLR) functionality was previously limited in the range of p-values it could report. Very small p-values were rounded to 0.0 due to floating-point precision limitations.

This change addresses the issue by increasing the default floating-point precision from `float32` to `float64`. This allows for the representation of much smaller numbers, significantly expanding the range of reportable p-values.

+++ tecpg/config.py
@@ -23,8 +23,8 @@ DEFAULT_DISTAL_DOWNSTREAM = 0  # No downstream search
 DEFAULT_DISTAL_UPSTREAM = 500_000_000  # To the end of the chromosome
 DEFAULT_DISTAL_WINDOW_BASE = 50_000  # 50 Kb
 
-DTYPE = torch.float32
-DEFAULT_FLOAT_FORMAT = '%.16f'
+DTYPE = torch.float64
+DEFAULT_FLOAT_FORMAT = '%.32f'
 
 
 def get_device(*, logger: Logger = Logger()) -> torch.device:
